Finished Implementing FetchedResultController

+++ SampleToDo/Controller/FolderViewController/FolderViewControllerViewModel.swift
@@ -11,14 +11,16 @@ import CoreData
 class FolderViewControllerViewModel {
     //CoreDataManager object should not be created in viewModel
     var coreDataManager: CoreDataManager = CoreDataManager()
-    var folderArray = [ToDoFolder]()
-
-    func getObject(at indexPath: IndexPath) -> ToDoFolder {
-        return folderArray[indexPath.row]
+//    var folderArray = [ToDoFolder]()
+    var fetchedResultController: NSFetchedResultsController<ToDoFolder>!
+    
+    init() {
+        fetchedResultController = setUpFetchedResultController()
     }
     
+    
     func swap(from source: Int, to destination: Int) {
-        self.folderArray.swapAt(source,destination)
+//        self.folderArray.swapAt(source,destination)
 //        coreDataManager.save()
     }
     
@@ -28,13 +30,29 @@ class FolderViewControllerViewModel {
         newFolder.creationDate = Date()
         newFolder.arrayOfList = []
         coreDataManager.save()
-        folderArray.append(newFolder)
+//        folderArray.append(newFolder)
     }
     
     func deleteFolder(at indexPath: IndexPath) {
-        let folderToBeDeleted = folderArray[indexPath.row]
-        coreDataManager.viewContext.delete(folderToBeDeleted)
-        folderArray.remove(at: indexPath.row)
+        if let folderToBeDeleted = fetchedResultController?.object(at: indexPath) {
+            coreDataManager.viewContext.delete(folderToBeDeleted)
+        }
+//        folderArray.remove(at: indexPath.row)
         coreDataManager.save()
+    }
+    
+    func setUpFetchedResultController() -> NSFetchedResultsController<ToDoFolder> {
+        let fetchRequest: NSFetchRequest<ToDoFolder> = ToDoFolder.fetchRequest()
+        let sort = NSSortDescriptor(key: "creationDate", ascending: true)
+        fetchRequest.sortDescriptors = [sort]
+        return NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext:coreDataManager.viewContext, sectionNameKeyPath: nil, cacheName: nil)
+    }
+    
+    func performFetch() {
+        do {
+            try fetchedResultController?.performFetch()
+        } catch let error {
+            fatalError("Error while initialising FetchedController. Error: \(error)")
+        }
     }
 }
